nomachine-enterprise-client 9.2.18_1
`nomachine-enterprise-client` is autobumped but it hasn't been updated
since version 8.16.1_2 because the `livecheck` block is broken. The
download page requires a cookie or the server will redirect to a login
page instead, which may be what's happening with the existing check.
Unlike `nomachine`, I haven't yet found a fix for this yet.

In the interim time, this updates the cask to the newest version after
manually checking the upstream download page.